Fixed useSearchParams issues on static pages by React suspense component. Fixed pre-rendering issues like undefined pages, undefined variables and functions. Updated .env.production in local, but passing the env variables from drployment directly in Vercel.
